Remove -mp from src/libs/sources/Makefile
This is from me testing configurations for the GPU porting of the MU solver. Removing the -mp from the Makefile in scr/sources for the actuator line should fix it. -Will fix it when the new AL version is ready for porting - For now the -mp can be removed.
I just compiled the code with intel and it compiles without errors.
The MPI implementation of MU needs revisiting, before being used for running test cases.